config: Remove unreachable memory error handling paths
None of these code paths could be reached (and most were invalid anyway,
as g_free() is the wrong destructor for collection_iterator).

g_strdup() will return NULL iff its input is NULL. On allocation
failure, it will abort the application, as per GLib policy. If you are
writing an application using GLib, attempting to handle ENOMEM
situations is pointless, but that’s a separate issue.

In all of these cases, the output of g_strdup() was being compared with
NULL on a code path dominated by a condition forcing the input to be
non-NULL.