sip: Simplify getting local IP address
Instead of iteraring the list of interfaces, skipping some and taking
the first one available, pick the IP address of the default route. This
is done by creating a UDP socket and connecting to address 1.2.3.4.
Because the socket is UDP no traffic is sent, but we can query the local
address using getsockname.