Integrate support for building JavaScript modules within the build system.
Building JavaScript modules is opt-in, i.e. disabled by default.
Various 'flavours' of JavaScript module syntax are supported:

 - ECMAScript (es)
 - CommonJS (cjs)
 - Immediately Invoked Function Expression (iife)
 - Universal Module Definition (umd)

To enable a particular flavour one of the --enable-javascript-*-module flags may be passed with an argument of 'yes' during ./configure:

 - --enable-javascript-es-module
 - --enable-javascript-common-js-module
 - --enable-javascript-iife-module
 - --enable-javascript-umd-module

Additionally, configure.ac is extended to lookup for an available version of npm if any of the supported module flavours is enabled.
If no npm can be found and a module flavour is enabled, ./configure will now fail with an error message accordingly.
Nota bene: this check for npm applies if and only if modules are actually enabled, therefore by default Janus will still build without npm.

When modules are enabled, generated modules will be built in ./npm/bundles (during make) and installed to $PREFIX/share/janus/javascript (during make install)
Additionally janus.js itself is now always installed to $PREFIX/share/janus/javascript as well (even if no modules are enabled).
The idea is that $PREFIX/share/janus/javascript acts as the installation location for the generically re-usable JavaScript libraries which are part of Janus.