Fix issue #1030 by introducing a new flag to videoroom for joining events
With private_id and require_pvtid, one can actually create a videoroom
with proper identity attached to every listener. However, there is no
event delivered when a listening only participant joins the room. This
requires the admin to have to poll the participants list to see if any
participants have joined and may have already been listening.
Added a new flag notify_joining to create API of the videoroom to allow
enabling an extra event sent when a participant joins the room. This
allows proper management of listening only participants by the admin.

Verified that with notify_joining enabled, a joining event is delivered
to all current participants with the new participants id and display.

Change-Id: I8217d112e65b1344d52f1c837ebd8b1f9657cd89