Add openssl pre 1.1 api macro
Instead of repeating the check for OPENSSL_VERSION_NUMBER < 0x10100000L or
defined(LIBRESSL_VERSION_NUMBER) over and over, it's cleaner to add the
checks for LIBRESSL_VERSION_NUMBER in dtls-bio.h and set a new define
JANUS_USE_OPENSSL_PRE_1_1_API if LibreSSL or OpenSSL < 1.1 are being
compiled against.

This will ease the transition to OpenSSL 1.0.2/1.1 support in the future
and will keep the code cleaner and easier to manage.