Remove unneeded allocations checking
Since commit 1f067658 ("Converted memory allocations to GLib ones, and
fixed a couple of leaks"), allocations are done with g_malloc() and
others. According to GLib documentation, these functions always return a
valid pointer, aborting the program when memory could not be allocated.
This effectively makes the checks useless, so remove them.